add last_n_rows_contents_formatted to Screen
This patch adds a new method to the Screen struct
which allows for re-drawing not just the current
screen, but a bit of additional scrollback context.
It might be possible to do this with the existing
API, but I couldn't figure out how after trying
a few hacky solutions (in particular, making an
the size extra big with set_size didn't seem to
work for some reason).