fix out of bounds cursor restore bug
This patch fixes an issue where the grid could get an
out of bounds position when a cursor_restore control
code is issued after the virtual terminal has been
resized to a smaller size since the position was saved.

I discovered this because starting or resuming vim saves
the cursor position and suspending it issues a cursor
restore control code (I'm not sure if it is vim or bash
generating these codes, bash would make more sense).

The fix I decided on is to just clamp the cursor to the
edge of the terminal in the event that this happens.